🐛 fix TypeScript Issues
